Chore: Update line-length rules in template files
We are losing too much time to the existing line length limits in our
linting setup. The AI coding tools struggle with this, and constantly
wrapping lines to 80 characters is wasting precious time/effort. Also,
it often introduces bugs/breakage that require additional rounds of
fixes. I've lost hours to this now, and now believe strongly that the
marginal benefits do not justify the existence of this rule.

I will happily go to war on this topic.

This pull request adjusts the limits to a more modern 120 characters.

Signed-off-by: ModeSevenIndustrialSolutions <mwatkins@linuxfoundation.org>